fix try-except and project structure
