+++ .docker/docker-compose.yml
@@ -20,9 +20,8 @@ services:
     image: mongo
     ports:
       - "30000:27017"
-    environment:
-      - MONGO_INITDB_ROOT_USERNAME=admin
-      - MONGO_INITDB_ROOT_PASSWORD=password
+    env_file:
+      - .env
     volumes:
       - mongo:/data/db
     networks:
