+++ report/routers/report_endpoints.py
@@ -2,14 +2,11 @@ from bson import ObjectId
 from fastapi import APIRouter, status
 from fastapi.responses import Response
 
-from report.routers.consumer import consumer
-from report.routers.database import (
+from report.models.database import (
     personal_workload_reports_collection,
     teachers_reports_collection,
 )
-from report.routers.exceptions import FeatureNotFindException
-from report.routers.generate_reports import generate_teachers_reports
-from report.routers.models import (
+from report.models.models import (
     ListPersonalWorkloadReport,
     ListTeacherReports,
     PersonalWorkloadReport,
@@ -17,6 +14,8 @@ from report.routers.models import (
     TeacherReport,
     TeacherReportCreateResponse,
 )
+from report.routers.exceptions import FeatureNotFindException
+from report.services.consumer import consumer
 
 router = APIRouter()
 
@@ -94,21 +93,6 @@ async def create_personal_workload_report(
         )
     )
     return created_personal_workload_report
-
-
-@router.post("/teacher-reports/create/", response_model=ListTeacherReports)
-async def create_teacher_report(reports: ListTeacherReports):
-    reports = await generate_teachers_reports()
-    created_reports = []
-    for report in reports:
-        await teachers_reports_collection.insert_one(
-            report.model_dump(by_alias=True, exclude=["id"])
-        )
-        created_report = await teachers_reports_collection.find_one(
-            {"_id": report.inserted_id}
-        )
-        created_reports.append(created_report)
-    return created_reports
 
 
 @router.get("consume/")
